+++ src/lib/company.spec.ts
@@ -1,5 +1,5 @@
 import { describe, test } from '@jest/globals';
-import { getCompany } from './company';
+import { getAllCompanies, getCompany, getSP500Companies } from '../index';
 
 const symbolsResponseText =
   '1\tABC\tABC Test Company Inc.\t9\t30\n1\tDEF\tDEF Test Company Inc.\t9\t122';
@@ -140,6 +140,8 @@ const transcriptLevel3ResponseJson = {
   ],
 };
 
+const sp500CompaniesTxtFile = 'ABC\nDEF';
+
 beforeAll(() => {
   global.fetch = jest.fn((url: string) => {
     if (url === 'https://v2.api.earningscall.biz/symbols-v2.txt?apikey=demo') {
@@ -192,6 +194,16 @@ beforeAll(() => {
       });
     }
 
+    if (
+      url === 'https://v2.api.earningscall.biz/symbols/sp500.txt?apikey=demo'
+    ) {
+      return Promise.resolve({
+        ok: true,
+        status: 200,
+        text: () => Promise.resolve(sp500CompaniesTxtFile),
+      });
+    }
+
     return Promise.reject(new Error(`Unhandled request: ${url}`));
   }) as jest.Mock;
 });
@@ -202,7 +214,7 @@ afterAll(() => {
 
 describe('company', () => {
   test('getCompany', async () => {
-    const company = await getCompany('ABC');
+    const company = await getCompany({ symbol: 'ABC' });
     expect(company.name).toBe('ABC Test Company Inc.');
     expect(company.companyInfo.exchange).toBe('NASDAQ');
     expect(company.companyInfo.symbol).toBe('ABC');
@@ -210,7 +222,7 @@ describe('company', () => {
     expect(company.companyInfo.industry).toBe('Consumer Electronics');
     expect(company.toString()).toBe('ABC Test Company Inc.');
 
-    const company2 = await getCompany('DEF');
+    const company2 = await getCompany({ symbol: 'DEF' });
     expect(company2.name).toBe('DEF Test Company Inc.');
     expect(company2.companyInfo.exchange).toBe('NASDAQ');
     expect(company2.companyInfo.symbol).toBe('DEF');
@@ -220,7 +232,7 @@ describe('company', () => {
   });
 
   test('get events', async () => {
-    const company = await getCompany('ABC');
+    const company = await getCompany({ symbol: 'ABC' });
     const events = await company.events();
     expect(events.length).toEqual(8);
     expect(events[0].year).toEqual(2024);
@@ -229,7 +241,7 @@ describe('company', () => {
   });
 
   test('get transcript', async () => {
-    const company = await getCompany('ABC');
+    const company = await getCompany({ symbol: 'ABC' });
 
     await expect(company.getTranscript({})).rejects.toThrow(
       'Must specify either event or year and quarter',
@@ -266,7 +278,7 @@ describe('company', () => {
   });
 
   test('get transcript level 2', async () => {
-    const company = await getCompany('ABC');
+    const company = await getCompany({ symbol: 'ABC' });
 
     const transcript = await company.getTranscript({
       year: 2022,
@@ -280,10 +292,14 @@ describe('company', () => {
     expect(transcript.text).toBe(
       "Good day and welcome to the ABC Test Company Inc. Q1 FY 2022 earnings conference call. Today's call is being recorded. Please go ahead. Thank you. Good afternoon, and thank you for joining us. I am the awesome CEO John Smith.",
     );
+    expect(transcript.speakers[0].speaker_info?.name).toBe('Operator');
+    expect(transcript.speakers[0].speaker_info?.title).toBe('Host');
+    expect(transcript.speakers[1].speaker_info?.name).toBe('John Smith');
+    expect(transcript.speakers[1].speaker_info?.title).toBe('CEO');
   });
 
   test('get transcript level 3', async () => {
-    const company = await getCompany('ABC');
+    const company = await getCompany({ symbol: 'ABC' });
 
     const transcript = await company.getTranscript({
       year: 2022,
@@ -297,5 +313,18 @@ describe('company', () => {
     expect(transcript.text).toBe(
       'Good day and welcome to the ABC Test Company Inc. Q1 FY 2022 earnings conference call. Thank you Good afternoon and thank you for joining us I am the awesome CEO John Smith.',
     );
+  });
+
+  test('getAllCompanies', async () => {
+    const companies = await getAllCompanies();
+    expect(companies.length).toBe(2);
+    const names = companies.map((company) => company.companyInfo.name);
+    expect(names).toContain('ABC Test Company Inc.');
+    expect(names).toContain('DEF Test Company Inc.');
+  });
+
+  test('getSP500Companies', async () => {
+    const companies = await getSP500Companies();
+    expect(companies.length).toBe(2);
   });
 });